Major changes to the directory structure
This commit changes the way JBackup looks for files.
Actions and rules are located under several data paths.
Which path gets saved to depends on user permissions.
Users with root privileges can write to the system
directory, and everyone else can write to their user
directory. An option is added to list the data paths.

This also adds options for listing available actions
and rules in all the data paths.

An option is added to list available logging levels.